refactor : 캔들 차트 이동 #75
